move utility functions out of project root
